infra/cleanup-function/requirements.txt
functions-framework==3.*
google-cloud-compute==1.*

